[new cases]: nearestCases Twilio Function now writes to db table NewCasesSubscriber, just need to make sure that actual phone number is being stored
