+++ lib/index.js
@@ -1,4 +1,5 @@
 var axios = require('axios')
+var mysql = require('mysql2')
 const opencage = require('opencage-api-client')
 const { messageTableDB, hospitalTableDB } = require('../constants')
 const haversineDistance = require('./haversineDistance')
@@ -145,11 +146,34 @@ const nearestCasesResponse = async (postalCode) => {
       response += `Total ${key}: ${value}.\n`
     }
 
-    return response
+    return response + '\nYou are now subscribed to alerts, to unsubscribe at any time text STOP'
   }
   else {
     return 'Unfortunately this service is only available in Ontario but we are working hard to onboard other provinces and will notify you when we are available.'
   }
+}
+
+const storeNewCasesSubscriber = (postalCode, number) => {
+  // Add a way to check that user is not in db table before adding them
+
+  const connection = mysql.createConnection({
+    host: process.env.MYSQL_HOST,
+    user: process.env.MYSQL_USERNAME,
+    password: process.env.MYSQL_PASS,
+    database: process.env.MYSQL_DATABASE,
+    port: process.env.MYSQL_PORT
+  })
+
+  let response = false
+
+  connection.query(`Insert Into NewCasesSubscribers (Subscriber, PostalCode, Active, createdAt, updatedAt)
+  Values (${number}, '${postalCode}', 1, NOW(), NOW())`, (error, results) => {
+    response = results && results.affectedRows > 0
+    if (error) console.log(error)
+    console.log('response:' + response)
+  })
+
+  return response
 }
 
 module.exports = {
@@ -160,5 +184,6 @@ module.exports = {
   setLanguageOptions,
   addNumberToGoogleSheet,
   getNearestCases,
-  nearestCasesResponse
+  nearestCasesResponse,
+  storeNewCasesSubscriber
 }
